@@ -31,7 +31,7 @@ from ..wsgi import get_path_info
 from ..wsgi import wrap_file
 
 _TOpener = t.Callable[[], tuple[t.IO[bytes], datetime, int]]
-_TLoader = t.Callable[[t.Optional[str]], tuple[t.Optional[str], t.Optional[_TOpener]]]
+_TLoader = t.Callable[[str | None], tuple[str | None, _TOpener | None]]
 
 if t.TYPE_CHECKING:
     from _typeshed.wsgi import StartResponse
